DLS-9710: Moved CBCSessionCache into new repositories package
